Move other pow types to units
