+++ units/Cargo.toml
@@ -13,15 +13,17 @@ rust-version = "1.63.0"
 exclude = ["tests", "contrib"]
 
 [features]
-default = ["std"]
+default = ["std", "hex"]
 std = ["alloc", "internals/std"]
-alloc = ["internals/alloc","serde?/alloc"]
+alloc = ["internals/alloc", "hex?/alloc", "serde?/alloc"]
+hex = ["dep:hex"]               # We don't anything from intenels/hex ATM.
 
 [dependencies]
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.4.0" }
 
-serde = { version = "1.0.103", default-features = false, optional = true }
 arbitrary = { version = "1.4", optional = true }
+hex = { package = "hex-conservative", version = "0.3.0", default-features = false, optional = true }
+serde = { version = "1.0.103", default-features = false, optional = true }
 
 [dev-dependencies]
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.4.0", features = ["test-serde"] }
